requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.4
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
phonenumbers==8.12.48
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1